+++ tapo/src/responses/device_info_result.rs
@@ -1,4 +1,5 @@
 mod color_light;
+mod default_plug_state;
 mod default_state;
 mod generic;
 mod hub;
@@ -11,6 +12,7 @@ mod rgb_light_strip;
 mod rgbic_light_strip;
 
 pub use color_light::*;
+pub use default_plug_state::*;
 pub use default_state::*;
 pub use generic::*;
 pub use hub::*;