+++ tapo/src/responses/device_info_result/default_plug_state.rs
@@ -1,0 +1,31 @@
+use serde::{Deserialize, Serialize};
+
+/// Plug Default State.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case", tag = "type")]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
+#[allow(missing_docs)]
+pub enum DefaultPlugState {
+    Custom { state: PlugState },
+    LastStates {},
+}
+
+/// Plug State.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
+#[allow(missing_docs)]
+pub struct PlugState {
+    pub on: bool,
+}
+
+#[cfg(feature = "python")]
+#[pyo3::pymethods]
+impl PlugState {
+    /// Gets all the properties of this result as a dictionary.
+    pub fn to_dict(&self, py: pyo3::Python) -> pyo3::PyResult<pyo3::Py<pyo3::types::PyDict>> {
+        let value = serde_json::to_value(self)
+            .map_err(|e| pyo3::exceptions::PyException::new_err(e.to_string()))?;
+
+        crate::python::serde_object_to_py_dict(py, &value)
+    }
+}